frontend/requirements.txt
flask
python-dotenv
requests
flask-login
PyJWT